feat: migrate traces to use daemonset
Adjusting the replica count using kustomize is a bit messy, so for now
we will migrate collection to be performed by a daemonset. We're still
in the process of adjusting resource limits for these manifests, but
this commit puts all the manifests in place required for starting to
play around with otel-collector.