feat(metrics): allow signalling pod should not be scraped (#36)
Under certain cases we collect more useful telemetry via logs than
metrics. For example, nginx-ingress generates a lot of metrics, all of
which are mostly useless once you are capable of ingesting the raw
access logs. This commit adds the ability to set a
`observeinc.com/scrape` annotation to a pod which behaves in the same
manner as `prometheus.io/scrape`, but targets observe collection
specifically.