feat: allow overriding environment variables with configMap (#38)
While it is relatively simple to override particular env variables using
kustomize, it is not very convenient if you're not well versed in the
tool. For quick adjustments that persist across updates, it can be
easier to just create a configMap of overrides rather than persist a
kustomization.yaml in a directory and then figure out how to keep track
of changes.

This change adds optional configMap overrides for each of our bases. We
allow for a specific, targeted configMap, or a global `env-overrides`
configMap. Unlike with kustomize, configuration changes won't be
automatically picked up, so a manual rollout is required.