feat(metrics): protect against cardinality explosion (#89)
Add configuration variables for limiting the number of samples and body
size per scrape. This is primarily intended to protect against
cardinality explosion. If a given target has an unbounded increase in
the number of samples, we should try to error out rather than run into
OOMs.

The configured bounds are conservative, but may still lead to false
positives, such as in the case of `kube-state-metrics` running in a very
large cluster. We happen to not need that data however since we collect
state data separately, so we end up erring on the side of caution.