fix(winlogs): pass token in query parameter (#106)
This commit changes the authentication method from Bearer token to query
parameter.

Some versions of Fluentd do not support interpolating environment
variables in the `headers` map.

We could instead use basic auth, but that would require us to split the
datatream token in the id (username) and the secret (password). This
cannot be easily done in either the configuration file or Kubernetes.

This leaves the query parameter option as the best remaining solution.
We already know string interpolation works correctly for string fields,
so we can attach the token there.