feat(traces): allow OBSERVE_CLUSTER override
Reintroduce the ability to use an `OBSERVE_CLUSTER` environment variable
to override the existing `k8s.cluster.uid` resource attribute. This can
be useful when migrating clusters.

In order for this override to take effect, the
`OTEL_OBSERVE_CLUSTER_ACTION` variable must be set to `upsert`. This is
a somewhat clumsy API but it works around limitations in the ability to
compose OTEL configuration in Kustomize.